backend/requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Werkzeug==2.3.6
pymongo==4.4.1
numpy==1.26.1
scikit-learn==1.3.2
python-dotenv==1.0.0
